MacOSX support: workspace changes
We no longer build glib from source, but instead require it to be
available locally.

Also modify other "local system" repositories to point at homebrew
directory locations.

This is not strictly better or worse than the master branch, since
users would generally be expected to adjust the local repositories
anyway, but at the same time this is not something we'll want to
integrate into master.

Also update the README.md to state the nature of this branch.